+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TypeTree 2.9-git
 -   Remove `apply_to_tuple` utility.
 -   Add support for `ProxyNode` on nodes with dynamic degree.
 -   The static member variable `CHILDREN` is deprecated in favor of the function `degree()`.
+-   Add comparison operator for `HybridTreePath`.
 
 TypeTree 2.8
 ------------